moved confetti and growing number to own file
