+++ apps/web/src/app/(wrapped)/wrapped/components/Text.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+import { useEffect, useState } from "react";
 import { motion } from "motion/react";
 
 export function AppearingText({ children, delay }: { children: React.ReactNode; delay: number }) {
@@ -35,3 +36,37 @@ export function InYourFace({ children, delay }: { children: React.ReactNode; del
     </motion.div>
   );
 }
+
+export const AnimatedNumber = ({ target }: { target: number }) => {
+  const [displayNumber, setDisplayNumber] = useState(0);
+  const animationDuration = 5; // in seconds
+
+  useEffect(() => {
+    const startTime = Date.now();
+    const updateNumber = () => {
+      const elapsedTime = (Date.now() - startTime) / 1000; // in seconds
+      const progress = Math.min(elapsedTime / animationDuration, 1);
+      const currentNumber = Math.floor(progress * target);
+      setDisplayNumber(currentNumber);
+
+      if (progress < 1) {
+        requestAnimationFrame(updateNumber);
+      }
+    };
+
+    updateNumber();
+  }, [target]);
+
+  return (
+    <motion.div
+      initial={{ fontSize: "1em", opacity: 0 }}
+      animate={{ fontSize: "13em", opacity: 1 }}
+      transition={{
+        duration: animationDuration,
+      }}
+      style={{ textAlign: "center", fontWeight: "bold" }}
+    >
+      {displayNumber}
+    </motion.div>
+  );
+};
